feat: complete star and premium pricing flow

+++ entity/starpackage.go
@@ -4,7 +4,7 @@ import "time"
 
 type StarPackage struct {
 	ID        uint64
-	Amount    uint32
+	Amount    uint64
 	Active    bool
 	CreatedAt time.Time
 	UpdatedAt time.Time
